fix(readme): report CI status for main and read the Rust version from Cargo.toml
The CI badge used the legacy /workflows/<name>/badge.svg URL, which is not
branch-scoped: it renders the newest run on any branch, so an in-flight pull
request made the README claim main was broken while main was green. Point it
at the workflow file with an explicit branch=main, and link to the same
filtered view.

The Rust badge hardcoded 1.70+ while Cargo.toml declares rust-version 1.88
(edition 2024, CI on 1.97.1) -- wrong for anyone checking what to install.
Read the value from the manifest so it cannot drift again.

The Docker Hub badge needed no change: "rate limited by upstream service" was
a transient shields.io throttle, and it resolves to v0.84.0 now. crates.io,
docs.rs and the licence badge all resolve correctly.

Both regressions are pinned by tests that fail against the previous README.
Adds crater parse/listing coverage and splits its test module into
src/crater_tests.rs to stay under the 1000-line limit.
